cart-pal: Change settings path to icarus/
Changed cart-pal to use icarus's settings instead of its own separate settings.

cart-pal differs from icarus in that it supports more iNES mappers and will generate manifests for the VS. System, PlayChoice-10, FamicomBox, and Atari 2600, but at the end of the day, it still uses the same settings as icarus.